requirements/requirements.txt
scikit-learn
pandas
PyYAML
fsspec
s3fs
scipy
seaborn
catboost
lightgbm
mlflow
boto3